+++ CHANGELOG.md
@@ -1,11 +1,12 @@
-﻿# Changelog
+# Changelog
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 ### Added
 - Project scaffold, CI, and version bump script.
 
-### Changed
+
+- add /events API and dashboard page### Changed
 - None
 
 ### Fixed